fix:modal & resource:ScheduleView
