+++ FreeSubject/ResetPage/ScheduleViewController.swift
@@ -1,0 +1,17 @@
+//
+//  ScheduleViewController.swift
+//  FreeSubject
+//
+//  Created by 홍준혁 on 2022/10/01.
+
+import UIKit
+import SnapKit
+import Then
+import RealmSwift
+import Realm
+
+class ScheduleViewController : UIViewController{
+    override func viewDidLoad() {
+        view.backgroundColor = .systemBackground
+    }
+}
